[Improvements]
Fix spawn of pin icon at 0 rather than correct position

Changed default taskbar order for new launch

Update install.sh

+++ install.sh
@@ -32,5 +32,7 @@ sudo cp winbar/winbar.cfg /etc
 sudo rm -r winbar
 
 # (old install location which should be deleted if still exists)
-sudo rm /usr/local/bin/winbar
+if [[ -f /usr/local/bin/winbar ]]; then
+  sudo rm /usr/local/bin/winbar
+fi
